Move snippets' get_queryset and paginate_queryset methods to generic IndexView
